debug

+++ src/com/company/dao/QuestionDao.java
@@ -5,6 +5,6 @@ import com.company.domain.Question;
 import java.math.BigDecimal;
 
 public interface QuestionDao {
-    public Question findByID(short questionNo,String testNo);
-    public void create(short questionNo, String testNo, String compulsory, String type, BigDecimal score, String content, String answer);
+    public Question findByID(int questionNo,String testNo);
+    public void create(int questionNo, String testNo, String compulsory, String type, BigDecimal score, String content, String answer);
 }
